requirement.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic
pydantic-settings==2.1.0
supabase==2.5.0
google-generativeai==0.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
asyncio==3.4.3
